requirements.txt
flask
pymongo[srv]
pandas
groq
python-dotenv
openpyxl